Separate form tabs to the new component. Join form and user info to the one component
