+++ src/front/shared/pages/Exchange/index.scss
@@ -97,7 +97,7 @@
   line-height: 1em;
 }
 
-.button.noMany {
+.button.noMoney {
   cursor: not-allowed;
   filter: grayscale(1);
   opacity: 0.15;
